Hide a button to create a new dashboard when adding private questions to a dashboard
