+++ frontend/src/metabase/containers/DashboardPicker.tsx
@@ -5,12 +5,16 @@ export interface DashboardPickerProps {
   value?: DashboardId;
   onChange: (dashboardId: DashboardId) => void;
   collectionId?: CollectionId;
+  shouldFetchDashboards?: boolean;
+  onOpenCollectionChange: (collectionId: CollectionId) => void;
 }
 
 const DashboardPicker = ({
   value,
   onChange,
   collectionId,
+  shouldFetchDashboards,
+  onOpenCollectionChange,
   ...props
 }: DashboardPickerProps) => (
   <ItemPicker
@@ -19,6 +23,8 @@ const DashboardPicker = ({
     value={value === undefined ? undefined : { model: "dashboard", id: value }}
     onChange={dashboard => onChange(dashboard.id)}
     models={["dashboard"]}
+    shouldFetchItems={shouldFetchDashboards}
+    onOpenCollectionChange={onOpenCollectionChange}
   />
 );
